+++ examples/cpp/demo1/demo1.cpp
@@ -33,8 +33,8 @@ int main()
       creg c[2];
       h q[0];
       x q[1];
-      measure q[1] -> c[1];
       measure q[0] -> c[0];
+      measure q[1] -> c[1];
     }
     )";