+++ tests/python_module/simple_test.py
@@ -129,46 +129,6 @@ def test_CI_210826_13_openqasm_index_out_of_range() :
     with pytest.raises((ValueError, RuntimeError)):
         s.run()
 
-def test_CI_210826_14_qbtheta_parameters() :
-    print("* CI_210826_14_qbtheta_parameters:")
-    print("* With default init settings, check the functionality for qbtheta parameter substitution")
-    import qristal.core
-    s = qristal.core.session()
-    s.init()
-    s.acc = "aer"
-    s.instring = '''
-    __qpu__ void qristal_circuit(qreg q) {
-        OPENQASM 2.0;
-        include "qelib1.inc";
-        creg c[1];
-        u3(QBTHETA_0, 1.6, QBTHETA_1) q[0];
-        measure q[0] -> c[0];
-    }'''
-    mth = qristal.core.MapIntDouble()
-    mth[0] = 0.05
-    mth[1] = -0.7
-    s.theta[0] = qristal.core.VectorMapIntDouble([mth])
-    s.run()
-
-    # Recompile the transpiled qasm to check
-    import xacc, math
-    compiler = xacc.getCompiler("staq")
-    ir = compiler.compile(s.out_transpiled_circuit[0][0]).getComposites()[0]
-    # u => 5 rx and ry gates
-    #  --Ry(pi/2)--Rx(lambda)--Ry(theta)--Rx(phi)--Ry(-pi/2)--
-    # 5 Rx/Ry gates + measure
-    assert(ir.nInstructions() == 6)
-    assert(ir.getInstruction(0).name() == "Ry")
-    assert(ir.getInstruction(0).getParameter(0) == pytest.approx(math.pi/2, abs=1e-3))
-    assert(ir.getInstruction(1).name() == "Rx")
-    assert(ir.getInstruction(1).getParameter(0) == pytest.approx(-0.7, abs=1e-3))
-    assert(ir.getInstruction(2).name() == "Ry")
-    assert(ir.getInstruction(2).getParameter(0) == pytest.approx(0.05, abs=1e-3))
-    assert(ir.getInstruction(3).name() == "Rx")
-    assert(ir.getInstruction(3).getParameter(0) == pytest.approx(1.6, abs=1e-3))
-    assert(ir.getInstruction(4).name() == "Ry")
-    assert(ir.getInstruction(4).getParameter(0) == pytest.approx(-math.pi/2, abs=1e-3))
-
 def test_raw_openqasm_str():
     print(" Testing raw OpenQASM string input ")
     import qristal.core, ast